requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
python-dotenv==1.0.0
pytest==7.4.3
pytest-flask==1.3.0
Flask-Testing==0.8.1
Flask-Analytics==0.6.0